Create build.sh
New build.sh - needed to be run before the main program can be used. It created the needed files and compiles main.c - Usage = sh build.sh <ROOT USERNAME> <ROOT PASSWORD> <SERVER IP> - username and password are what you want to use when rooting any server. The server IP is for a reverse shell to be created if you want.